pyproject.toml
[project]
name = "evals"
version = "0.1.1"
requires-python = ">=3.9"
dependencies = [
    "mypy",
    "openai >= 0.27.2",
    "tiktoken",
    "blobfile",
    "backoff",
    "numpy",
    "snowflake-connector-python[pandas]",
    "pandas",
    "fire",
    "pydantic",
    "tqdm",
    "nltk",
    "filelock",
    "mock",
    "langdetect",
    'termcolor',
    "lz4",
    "pyzstd",
    "pyyaml",
    "sacrebleu",
    "matplotlib",
    "datasets",
    "openpyxl",
    "rouge_score_ar @ git+https://github.com/ARBML/rouge_score_ar.git"
]

[project.scripts]
oaieval = "evals.cli.oaieval:main"
oaievalset = "evals.cli.oaievalset:main"
